chore(view): move Filter code to TemporalFilter
