--- packages/view/src/components/TemporalFilter/Filter/Filter.scss
@@ -1,16 +1,0 @@
-.filter {
-  font-weight: 600;
-  font-size: 0.825rem;
-
-  .date-from,
-  .date-to {
-    margin: 0 12px 0 8px;
-    padding: 2px;
-    border: solid 1px rgba(255, 255, 255, 0.22);
-    border-radius: 3px;
-    color: white;
-    background: rgba(255, 255, 255, 0.09);
-    font-weight: 300;
-    font-size: 0.825rem;
-  }
-}
