--- packages/view/src/components/TemporalFilter/Filter/index.ts
@@ -1,1 +1,0 @@
-export { default as Filter } from "./Filter";
