+++ packages/view/src/components/TemporalFilter/TemporalFilter.scss
@@ -27,8 +27,25 @@
   .line-chart {
     display: flex;
     flex-direction: column;
-    height:120px;
+    height: 120px;
     width: 100%;
     align-content: space-around;
+  }
+}
+
+.filter {
+  font-weight: 600;
+  font-size: 0.825rem;
+
+  .date-from,
+  .date-to {
+    margin: 0 12px 0 8px;
+    padding: 2px;
+    border: solid 1px rgba(255, 255, 255, 0.22);
+    border-radius: 3px;
+    color: white;
+    background: rgba(255, 255, 255, 0.09);
+    font-weight: 300;
+    font-size: 0.825rem;
   }
 }
